ml/requirements.txt
openai==1.3.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
supabase==2.3.0
python-dotenv==1.0.0
joblib==1.3.2
xgboost==2.0.3

